Cleaned up the 7 day tracker
