+++ client/src/pages/MobileLayout.tsx
@@ -4,13 +4,8 @@ import './MobileLayout.css';
 import { Link } from 'react-router-dom';
 import { LayoutProps } from './Types';
 import { Streaks } from './Streaks';
-import { useEffect, useState, useCallback } from 'react';
-import { getAuthHeaders } from '../lib/auth';
-
-type WeeklyTracker = {
-  emojiPath: string;
-  logDate: string;
-};
+import { useState } from 'react';
+import { WeeklyMoodTracker } from '../components/WeeklyMoodTracker';
 
 export function MobileLayout({
   isOpen,
@@ -25,58 +20,12 @@ export function MobileLayout({
   closeModal = () => {},
   user,
 }: LayoutProps) {
-  const [historyTracker, setHistoryTracker] = useState<WeeklyTracker[]>([]);
   const [refreshData, setRefreshData] = useState(0);
-
-  const fetchMoodHistory = useCallback(async () => {
-    try {
-      if (!user?.id) {
-        console.error('Failed to get user.id');
-        return;
-      }
-
-      const res = await fetch(`/api/mood-tracking/${user.id}/weekly`, {
-        headers: {
-          ...getAuthHeaders(),
-        },
-      });
-
-      if (!res.ok) throw new Error(`Response Status: ${res.status}`);
-
-      const data = await res.json();
-      setHistoryTracker(data);
-    } catch (err) {
-      console.error(`Error fetching weekly mood data: ${err}`);
-    }
-  }, [user?.id]);
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     await originalHandleSubmit(event);
     setRefreshData((prev) => prev + 1);
   };
-
-  useEffect(() => {
-    fetchMoodHistory();
-  }, [fetchMoodHistory, refreshData]);
-
-  const generatePastWeek = (): { date: string; abbr: string }[] => {
-    const result = [];
-    const today = new Date();
-
-    for (let i = 6; i >= 0; i--) {
-      const day = new Date();
-      day.setDate(today.getDate() - i);
-
-      result.push({
-        date: day.toISOString().split('T')[0],
-        abbr: ['S', 'M', 'T', 'W', 'T', 'F', 'S'][day.getDay()],
-      });
-    }
-
-    return result;
-  };
-
-  const pastWeek = generatePastWeek();
 
   return (
     <>
@@ -104,28 +53,7 @@ export function MobileLayout({
           />
 
           <div className="dashboard-row w-full h-15 items-center">
-            {pastWeek.map((day) => {
-              const dayMood = historyTracker.find(
-                (mood) =>
-                  new Date(mood.logDate).toISOString().split('T')[0] ===
-                  day.date
-              );
-
-              return (
-                <div key={day.date} className="flex flex-col items-center">
-                  <img
-                    className="emoji-icon"
-                    src={
-                      dayMood
-                        ? dayMood.emojiPath
-                        : '/images/emoji/gray-neutral.svg'
-                    }
-                    alt="mood"
-                  />
-                  <p>{day.abbr}</p>
-                </div>
-              );
-            })}
+            <WeeklyMoodTracker userId={user?.id} refreshTrigger={refreshData} />
           </div>
 
           <div className="dashboard-row center">
